feat: add more conditionality
