+++ index.ts
@@ -7,7 +7,7 @@ import "./resources/kubernetes/api/api.js";
 import "./resources/kubernetes/portal-api/redis.js";
 import "./resources/kubernetes/portal-api/portal-api.js";
 import "./resources/kubernetes/portal-app/portal-app.js";
-import "./resources/kubernetes/portal-app/portal-domains.js";
+import "./resources/kubernetes/customer-dns.js";
 import "./resources/kubernetes/registration-app/registration-app.js";
 import "./resources/kubernetes/registration-app/sanity-credentials.js";
 import "./resources/shared/shared-resources.js";
